pyproject.toml
[tool.poetry]
name = "watch-converter"
version = "0.1.0"
description = "Image converter for Apple Watch screen"
authors = ["asovchar <a.s.ovchar@gmail.com>"]
readme = "README.md"
repository = "https://github.com/asovchar/watch-converter"
packages = [{include = "converter"}]

[tool.poetry.dependencies]
python = "^3.8"
numba = "^0.51.0"
numpy = "^1.16.2"
pillow = "^6.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.2"
pre-commit = "^2.4.0"
pytest = "^6.0.1"
pytest-cov = "^2.10.0"

[tool.poetry.scripts]
converter = 'converter.__main__:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
